Fix typo in pagination param to "after_id" and moved `FeedEndpoint` to its own file in "Feed API" module
