+++ EssentialFeedTests/Feed API/FeedEndpointTests.swift
@@ -1,25 +1,6 @@
 import XCTest
 import EssentialFeed
 
-enum FeedEndpoint {
-    case get(after: FeedImage?)
-
-    func url(baseURL: URL) -> URL {
-        switch self {
-        case let .get(feedImage):
-            var component = URLComponents()
-            component.scheme = baseURL.scheme
-            component.host = baseURL.host
-            component.path = baseURL.path + "/v1/feed"
-            component.queryItems = [
-                feedImage.map { _ in URLQueryItem(name: "after", value: feedImage?.id.uuidString) },
-                URLQueryItem(name: "limit", value: "10")
-            ].compactMap { $0 }
-            return component.url!
-        }
-    }
-}
-
 class FeedEndpointTests: XCTestCase {
     func test_get_deliversFeedEndpoint() {
         let baseURL = URL(string: "http://base-url.com")!
@@ -36,6 +17,6 @@ class FeedEndpointTests: XCTestCase {
         let baseURL = URL(string: "http://base-url.com")!
         let received = FeedEndpoint.get(after: image).url(baseURL: baseURL)
 
-        XCTAssertEqual(received.query?.contains("after=\(image.id.uuidString)"), true, "after")
+        XCTAssertEqual(received.query?.contains("after_id=\(image.id.uuidString)"), true, "after")
     }
 }
